move account destruction to its own file
